+++ CHANGES.md
@@ -8,14 +8,16 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 
 @dev
    * Changed how auto-converter checks for supported classes (see [Data Conversion](docs/docs.md#data-conversion))
+   * Data conversion now supports [JsonResource](https://laravel.com/docs/6.0/eloquent-resources) data class.
    * Added unit test for `ResponseBuilderServiceProvider::mergeConfg()`.
+   * Moved data conversion code to separate `Converter` class.
 
 * 6.1.2 (2019-10-02)
    * Corrected ServiceProvider used for tests.
 
 * 6.1.1 (2019-10-02)
    * Fixed `ResponseBuilderServiceProvider` using unreferenced `Arr` class method (reported by yassir3wad).
-
+   
 * 6.1.0 (2019-09-28)
    * **BACKWARD INCOMPATIBLE CHANGES** ([more info](docs/compatibility.md))
    * Created new library logo (see [artwork/](artwork/) folder).